make rvc module, add rvc base model download
