+++ modules/__init__.py
@@ -3,3 +3,4 @@ import modules.reference_save
 import modules.reference_speaker
 import modules.select_models
 import modules.settings
+import modules.rvc
